+++ AeSys/EoGeUniquePoint.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#define btest(m, p) ((bool) (((((DWORD) m) >> ((int) p)) & 1UL) == 1 ? true : false))
+#define btest(m, p) ((bool) (((((unsigned long) m) >> ((int) p)) & 1UL) == 1 ? true : false))
 
 class EoGeUniquePoint : public CObject {
 public:
